% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lca_parameters_from_xls.R
\name{read_lca_parameters_from_xls}
\alias{read_lca_parameters_from_xls}
\title{Read LCA Parameters from an Excel File}
\usage{
read_lca_parameters_from_xls(file, country = "de")
}
\arguments{
\item{file}{path to the Excel file to be read}

\item{country}{country code: "de" for German or "en" for "English". Required
to convert the text values in the Value columns}
}
\description{
Read LCA Parameters from an Excel File
}
