% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lca_parameters_to_xlsx.R
\name{save_lca_parameters_to_xlsx}
\alias{save_lca_parameters_to_xlsx}
\title{Save LCA parameters to an Excel File}
\usage{
save_lca_parameters_to_xlsx(lca, file, overwrite = FALSE)
}
\arguments{
\item{lca}{data frame with lca parameters as returned by
\code{\link{read_lca_parameters_from_xls}}}

\item{file}{path to Excel file to be written}

\item{overwrite}{it \code{TRUE} an existing file will be overwritten, passed
to \code{\link[openxlsx]{saveWorkbook}}}
}
\description{
Save LCA parameters to an Excel File
}
